+++ mullvad-daemon/src/management_interface.rs
@@ -23,6 +23,7 @@ use std::sync::{Arc, Mutex, RwLock};
 use std::sync::atomic::{AtomicBool, Ordering};
 use talpid_core::mpsc::IntoSender;
 use talpid_ipc;
+use talpid_types::net::TunnelOptions;
 use uuid;
 
 use account_history::AccountHistory;
@@ -115,6 +116,14 @@ build_rpc_trait! {
         #[rpc(meta, name = "remove_account_from_history")]
         fn remove_account_from_history(&self, Self::Metadata, AccountToken) -> BoxFuture<(), Error>;
 
+        /// Sets openvpn's mssfix parameter
+        #[rpc(meta, name = "set_openvpn_mssfix")]
+        fn set_openvpn_mssfix(&self, Self::Metadata, Option<u16>) -> BoxFuture<(), Error>;
+
+        /// Gets tunnel specific options
+        #[rpc(meta, name = "get_tunnel_options")]
+        fn get_tunnel_options(&self, Self::Metadata) -> BoxFuture<TunnelOptions, Error>;
+
         #[pubsub(name = "new_state")] {
             /// Subscribes to the `new_state` event notifications.
             #[rpc(name = "new_state_subscribe")]
@@ -165,6 +174,10 @@ pub enum TunnelCommand {
     SetAllowLan(OneshotSender<()>, bool),
     /// Request the current allow LAN setting.
     GetAllowLan(OneshotSender<bool>),
+    /// Set the mssfix argument for OpenVPN
+    SetOpenVpnMssfix(OneshotSender<()>, Option<u16>),
+    /// Get the mssfix argument for OpenVPN
+    GetTunnelOptions(OneshotSender<TunnelOptions>),
     /// Makes the daemon exit the main loop and quit.
     Shutdown,
 }
@@ -553,6 +566,29 @@ impl<T: From<TunnelCommand> + 'static + Send> ManagementInterfaceApi for Managem
                     Error::internal_error()
                 }),
         ))
+    }
+
+    fn set_openvpn_mssfix(
+        &self,
+        meta: Self::Metadata,
+        mssfix: Option<u16>,
+    ) -> BoxFuture<(), Error> {
+        trace!("set_openvpn_mssfix");
+        try_future!(self.check_auth(&meta));
+        let (tx, rx) = sync::oneshot::channel();
+        let future = self.send_command_to_daemon(TunnelCommand::SetOpenVpnMssfix(tx, mssfix))
+            .and_then(|_| rx.map_err(|_| Error::internal_error()));
+
+        Box::new(future)
+    }
+
+    fn get_tunnel_options(&self, meta: Self::Metadata) -> BoxFuture<TunnelOptions, Error> {
+        trace!("get_tunnel_options");
+        try_future!(self.check_auth(&meta));
+        let (tx, rx) = sync::oneshot::channel();
+        let future = self.send_command_to_daemon(TunnelCommand::GetTunnelOptions(tx))
+            .and_then(|_| rx.map_err(|_| Error::internal_error()));
+        Box::new(future)
     }
 
     fn new_state_subscribe(